modify parseFhirSearch to handle an array of responses
